backend/requirements.txt
Django>=4.2
supabase>=2.0
PyJWT>=2.8
django-cors-headers>=4.3
gunicorn
